ContactARLab/app/src/main/java/com/contactar/contactarlaboratory/export/csv/WifiCsvFileWriter.java
package com.contactar.contactarlaboratory.export.csv;

import android.content.Context;

import com.contactar.contactarlaboratory.constants.SourceTypeEnum;
import com.contactar.contactarlaboratory.database.AppDatabase;
import com.contactar.contactarlaboratory.repositories.WifiRepository;

import java.util.ArrayList;
import java.util.List;

public class WifiCsvFileWriter extends CsvSampleFileWriter {

    public WifiCsvFileWriter(long[] runIds, AppDatabase database, Context context) throws NullPointerException {
        super(runIds, database, context);
    }

    @Override
    public List<CsvExportable> getExportableData() {
        return new ArrayList<>(new WifiRepository(database).getAllSamples(runIds));
    }

    @Override
    public String getKey() {
        return SourceTypeEnum.WIFI.name();
    }
}
